+++ internal/graphql/schema/where.go
@@ -297,6 +297,8 @@ func parseOperator(op string) (repositories.FilterOperator, bool) {
 		return "", true
 	case "eq":
 		return repositories.OpEq, false
+	case "eqi":
+		return repositories.OpEqi, false
 	case "neq":
 		return repositories.OpNeq, false
 	case "gt":
@@ -309,8 +311,6 @@ func parseOperator(op string) (repositories.FilterOperator, bool) {
 		return repositories.OpLte, false
 	case "in":
 		return repositories.OpIn, false
-	case "eqi":
-		return repositories.OpEqi, false
 	case "ini":
 		return repositories.OpIni, false
 	case "contains":